requirements.txt
numpy
albumentations
Cython
h5py
labelme
matplotlib
onnx
onnx-simplifier
onnxruntime
opencv-python
pandas
Pillow
scikit-learn
seaborn
sheen
tensorboard
timm
tqdm
visdom
